[tests/tcan_round.c] Added missing casts for printf (error messages). Bug detected under MS Windows, where size_t > unsigned long. See:   https://sympa.inria.fr/sympa/arc/mpfr/2018-08/msg00002.html
[From SVN r12983 (trunk)]

+++ tests/tcan_round.c
@@ -93,7 +93,7 @@ check_round_p (void)
         {
           printf ("mpfr_round_p(%d) != mpfr_can_round(%d,RNDZ)!\n"
                   "bn = %ld, err0 = %ld, prec = %lu\nbp = ",
-                  r1, r2, n, (long) err, (unsigned long) p);
+                  r1, r2, (long) n, (long) err, (unsigned long) p);
 #ifndef MPFR_USE_MINI_GMP
           gmp_printf ("%NX\n", buf, n);
 #endif
@@ -108,7 +108,7 @@ check_round_p (void)
         {
           printf ("mpfr_can_round(%d,RNDN) != mpfr_can_round(%d,RNDF)!\n"
                   "bn = %ld, err0 = %ld, prec = %lu\nbp = ",
-                  r1, r2, n, (long) err, (unsigned long) p);
+                  r1, r2, (long) n, (long) err, (unsigned long) p);
 #ifndef MPFR_USE_MINI_GMP
           gmp_printf ("%NX\n", buf, n);
 #endif
@@ -123,7 +123,7 @@ check_round_p (void)
         {
           printf ("mpfr_can_round(%d,RNDA) != mpfr_can_round(%d,RNDF)!\n"
                   "bn = %ld, err0 = %ld, prec = %lu\nbp = ",
-                  r1, r2, n, (long) err, (unsigned long) p);
+                  r1, r2, (long) n, (long) err, (unsigned long) p);
 #ifndef MPFR_USE_MINI_GMP
           gmp_printf ("%NX\n", buf, n);
 #endif
@@ -138,7 +138,7 @@ check_round_p (void)
         {
           printf ("mpfr_can_round(%d,RNDZ) != mpfr_can_round(%d,RNDF)!\n"
                   "bn = %ld, err0 = %ld, prec = %lu\nbp = ",
-                  r1, r2, n, (long) err, (unsigned long) p);
+                  r1, r2, (long) n, (long) err, (unsigned long) p);
 #ifndef MPFR_USE_MINI_GMP
           gmp_printf ("%NX\n", buf, n);
 #endif
